move more pileup-creation code out of TestUtil
